WIP: Upload medline and PMC OA and Manuscripts into database.
